Move CompressiveEncoder into compression module
