+++ src/prfmodel/models/__init__.py
@@ -3,6 +3,7 @@
 from .cf.canonical import CanonicalCFModel
 from .cf.gaussian import GaussianCFModel
 from .cf.stimulus_encoding import CFStimulusEncoder
+from .compression import CompressiveEncoder
 from .prf.canonical import CanonicalPRFModel
 from .prf.css import Gaussian2DCSSPRFModel
 from .prf.css import init_css_from_gaussian
@@ -13,7 +14,6 @@ from .prf.div_norm import init_dn_from_gaussian
 from .prf.dog import DoG2DPRFModel
 from .prf.dog import init_dog_from_gaussian
 from .prf.gaussian import Gaussian2DPRFModel
-from .prf.stimulus_encoding import CompressiveEncoder
 from .prf.stimulus_encoding import PRFStimulusEncoder
 from .prf.stimulus_encoding import encode_prf_response
